test/requirements.txt
numpy
opencv-python
torch
torchvision
torchaudio
pytz
Pillow
facenet-pytorch
faiss-cpu